CLN refactorize util+some clean up
