+++ benchopt/tests/test_cli.py
@@ -1,3 +1,4 @@
+import click
 import pytest
 
 
@@ -5,11 +6,11 @@ from benchopt.cli import run
 
 
 def test_invalid_benchmark():
-    with pytest.raises(SystemExit, match=r"2"):
-        run(['invalid_benchmark'], 'benchopt')
+    with pytest.raises(click.BadParameter, match=r"invalid_benchmark"):
+        run(['invalid_benchmark'], 'benchopt', standalone_mode=False)
 
 
 def test_invalid_dataset():
-    with pytest.raises(AssertionError, match=r"invalid_dataset"):
+    with pytest.raises(click.BadParameter, match=r"invalid_dataset"):
         run(['lasso', '-l', '-d', 'invalid_dataset', '-s', 'baseline'],
-            'benchopt')
+            'benchopt', standalone_mode=False)
